chore: TokenProviderTest 패키지 변경
